new features[menu, maze, navigation types]

+++ xclient.c
@@ -1,3 +1,4 @@
+
 /* FEI STU, predmet RTOS
  * Semestralne zadanie
  *
@@ -22,6 +23,7 @@
 #include "xtimer.h"
 #include "xpacket.h"
 #include "xsharedmemory.h"
+#include "xnavigation.h"
 
 //::Function declarations
 static void emptyFcn();
@@ -30,37 +32,60 @@ static void sendLidarData();
 static void *receiveFromServer();
 static void initRover();
 static void nextMove(int, int);
-static void updateRover(uint8_t);
+static void updateRover(uint8_t *);
+static void updateDeltaMove(uint8_t *);
 static void initScreen();
 static void repaintMap();
 static void setServerPID(uint8_t *);
 static void stopHandler();
 static void sendProcessID();
+static void sendCoordinates();
+static void sendSimplePacket(uint8_t );
+static void runServer();
 
 //::Global variables
 socketArgs_t client;
+coordinates_t coords;
 Point rover;
 pid_t serverPID;
+uint8_t *sharedTask;
 WINDOW *mapa = NULL;
 
 //::Main
 int main(int argc, char *argv[]) {
 	pthread_t thread1;
-	char *sharedTask;
+
 
 	//INITIALIZATION..
 	signal(SIGINT, stopHandler);
-	initClientSocket(&client);
-	createThread(&thread1, receiveFromServer, &client);
-	getSharedMemory(sharedTask, 32, 5678);
-	sendProcessID();
-
 	initSpace(argc,argv);
-	kill(serverPID, SIGUSR2); //Tell server we are ready, must be below initSpace()
+
+	/* Create Server interaction only if needed */
+	if(navigation.local.lidar.type) {
+		createProcess(runServer, emptyFcn, NULL, NULL);
+		/* Comunication with server*/
+		printf("\nCreating server process..\n");
+		sleep(3);
+
+		initClientSocket(&client);
+		createThread(&thread1, receiveFromServer, &client);
+		sendProcessID();
+		printf("Socket comunication available...\n");
+
+		sleep(1);
+		sharedTask = getSharedMemory(sizeof(navigation_t), 5678);
+		memset(sharedTask, 0, sizeof(navigation_t));
+		memcpy(sharedTask, &navigation, sizeof(navigation_t));
+		printf("Shared memory available..\n");
+
+		sleep(1);
+		printf("Server ready!");
+		kill(serverPID, SIGUSR2); //Tell server we are ready, must be below initSpace()
+	}
+
 	initScreen();
 	initRover();
 
-//	newProcess(socketProcessFcn,emptyFcn,(void *)&rover, NULL);
 
 	//CYCLING...
 	while(1) sleep(1);
@@ -71,12 +96,19 @@ int main(int argc, char *argv[]) {
 	return 0;
 }
 
+static void runServer()
+{
+
+//	system("xterm -x bash -c './server'; 'geometry = 50x50 10+10'; exec bash");
+	system("xterm -e './server' -- 'geometry=50x50'");
+//	system("xterm -e bash -c '--geometry=50x50+100+120; ./server; exec bash'");
+}
+
 static void stopHandler()
 {
 
 	close(client.initSocket_fd);
         close(client.sharedSocket_fd);
-
 	endwin();
 	printf("User wish ext program!\n");
 
@@ -99,10 +131,10 @@ static void initScreen()
 
 static void repaintMap()
 {
-        for(int i=0; i < SPACE_SIZE; i++){
-                for(int j=0; j < SPACE_SIZE*2; j++)
+        for(int y=0; y < SPACE_SIZE; y++){
+                for(int x=0; x < SPACE_SIZE*2; x++)
                 {
-                        mvwaddch(mapa,i, j,Space[i][j]);
+                        mvwaddch(mapa, y, x, Space[y][x]);
                 }
                 waddch(mapa, '\n');
         }
@@ -134,8 +166,9 @@ static void *receiveFromServer(socketArgs_t *args)
                         	if(p.crc == calcCRC)
                                 //PARSING PACKETS
                                 switch(p.header.type) {
-					case packet_type_move : updateRover(*p.data); break;
+					case packet_type_move : updateRover(p.data); break;
                                 	case packet_type_pid  : setServerPID(p.data); break;
+					case packet_type_delta_move : updateDeltaMove(p.data); break;
 				}
 
 				free(p.data);   //Uvolinit pamat
@@ -143,6 +176,7 @@ static void *receiveFromServer(socketArgs_t *args)
 			else {
 				switch(p.header.type) {
                                         case packet_type_lidar_data: sendLidarData(); break;
+					case packet_type_coordinates : sendCoordinates(); break;
                                 }
 			}  //PACKET WITHOUT DATA
                 } //PACKET MARK NOT FOUND
@@ -159,17 +193,25 @@ static void setServerPID(uint8_t *data)
 	serverPID = pID;
 }
 
-static void updateRover(uint8_t move)
+static void updateDeltaMove(uint8_t *data)
 {
-	switch(move) {
+	int16_t dx = ntohs(data[0] | data[1] << 8);
+	int16_t dy = ntohs(data[2] | data[3] << 8);
 
-		case up : nextMove(-1, 0); break;
-		case down : nextMove(1, 0); break;
-		case right : nextMove(0, 1); break;
-		case left : nextMove(0, -1); break;
+	nextMove(dx, dy);
+
+}
+
+static void updateRover(uint8_t *move)
+{
+	switch(*move) {
+
+		case up : nextMove(0, -1); break;
+		case down : nextMove(0, 1); break;
+		case right : nextMove(1, 0); break;
+		case left : nextMove(-1, 0); break;
 		default : printf("Invalid movement command!\n");
 	}
-	updateLidarData(rover);
 }
 
 static void nextMove(int x, int y)
@@ -181,28 +223,52 @@ static void nextMove(int x, int y)
 	if(isFree(next)) {
 		rover.x += x;
 		rover.y += y;
-		mvwaddch(mapa, rover.x, rover.y, '*');
+		mvwaddch(mapa, rover.y, rover.x, '*');
 		wrefresh(mapa);
 	}
+	updateLidarData(rover);
 }
 
 
 static void emptyFcn(){};
 
+static void sendCoordinates()
+{
+
+	coords.start = startPoint;
+	coords.actual = rover;
+	coords.goal = goalPoint;
+/*
+	printf("goalx :%d goaly: %d\n", coords.goal.x, coords.goal.y);
+        printf("actualx :%d actualy : %d\n", coords.actual.x, coords.actual.y);
+
+*/	Data_Packet p;
+        p.header.mark          = PACKET_MARK;
+        p.header.data_len      = sizeof(coordinates_t);
+        p.header.type          = packet_type_coordinates;
+        p.crc                  = calc_crc8((uint8_t *)&coords, p.header.data_len);
+
+        sendSocket(&client.initSocket_fd, (char *)&p, sizeof(Data_Packet) -1);
+        sendSocket(&client.initSocket_fd, (char *)&coords, p.header.data_len);
+
+	//GET NAVIGATION
+	sendSimplePacket(packet_type_navigation);
+}
+
 static void sendProcessID()
 {
-	      	uint16_t packet_size = sizeof(Data_Packet);
-        	Data_Packet  p;
-        	pid_t processID= getpid();
-        	uint8_t *ptr = (uint8_t *)&processID;
+     	uint16_t packet_size = sizeof(Data_Packet);
+       	Data_Packet  p;
+       	pid_t processID= getpid();
+       	uint8_t *ptr = (uint8_t *)&processID;
 
-        	p.header.mark          = PACKET_MARK;
-        	p.header.data_len      = sizeof(pid_t);
-        	p.header.type          = packet_type_pid;
-        	p.crc                  = calc_crc8((uint8_t *)ptr, p.header.data_len);
-        	p.data                 = ptr;
-        	sendSocket(&client.initSocket_fd, (char *)&p, packet_size -1);
-        	sendSocket(&client.initSocket_fd, (char *)p.data, sizeof(pid_t));
+       	p.header.mark          = PACKET_MARK;
+       	p.header.data_len      = sizeof(pid_t);
+       	p.header.type          = packet_type_pid;
+       	p.crc                  = calc_crc8((uint8_t *)ptr, p.header.data_len);
+       	p.data                 = ptr;
+       	sendSocket(&client.initSocket_fd, (char *)&p, packet_size -1);
+       	sendSocket(&client.initSocket_fd, (char *)p.data, sizeof(pid_t));
 }
 
 static void sendLidarData()
@@ -255,7 +321,7 @@ static void initClientSocket(socketArgs_t *args)
 
 static void initRover()
 {
-	rover.x = start.x+1;
-	rover.y = start.y+1;
+	rover.x = startPoint.x+1;
+	rover.y = startPoint.y+1;
 	updateLidarData(rover);
 }
